Fix bug in process_params extraction in followup pipeline
The logic previously implemented in the method switch_ifo() to extract the process_param table of the inspiral job run to generate a given trigger did not guarantee that the correct table was found. This was due to the fact that the trigger's time was compared to the gps-start-time, gps-end-time of the process_param table, but there can exist multiple tables matching these conditions. This happens for data chunks close to the end of the science segment, in this case two data chunks can overlap.
The correct way to guarantee that we select the data chunk in which the candidate was found is to compare the candidate's time to the search_summary.out_start_time and search_summary.out_end_time.