fix livetime bug
the livetime was being computed as the total time for which all available
instruments were on --- if a burca job spanned a time when only H1 and L1
were on, then the livetime was all time when exactly H1 and L1 were on, if
the burca job spanned H1+L1+V1 time then the livetime was the time when
exactly those three were on, basically full n-way coincidence where n is
however many instruments are in the file.  now livetime is all time when
at least 2 instruments are on.

this also adds support for ligolw_burca's new --coincidence-segment option
in the livetime algorithm:  if that option is found in the process_params
table, then the segments in the file are clipped to that interval during
the livetime summation.