remove SCALE factor from lalapps_StringSearch
this was previously in place to prevent underflows when measuring the PSD
when the code used single-precision arithmetic.  it is no longer required,
and removing it means that a "magic" number doesn't need to be duplicated
in several places in the pipeline.

this patch also addes some comments to the bisection search code used to
place templates, and fixes a memory leak therein.