stringutils:  fix veto application in likelihood data extraction
Instead of calling the .get_peak() method of each trigger to retrieve its
peak time and then check to see if that time is in the segmentlist, the
conditional was checking to see if the .get_peak method object itself was
in the segmentlist.  Not sure why this didn't produce a TypeError, but it
didn't.