report errors that occur in threads
when an exception occurs in a thread it produces a traceback on stderr, but
only the one thread terminates, the program keeps going.  this patch wraps
the function executed in the threads in a catch-all try/except block that
stores the exception info in a queue, allowing the main thread to report
them (and crash, as desired) if the threads fail.