lalapps_string_cs_gamma:  rework
- remove loop over P values because (a) plotting code will only accept P=1
anyway and (b) because the loop is a little screwed up if more than 1 value
is requested

- replace command line options --log-whatever with --whatever and user must
enter the actual number instead of the log-base-10 of the value.  this
clarifies what number should be entered because it wasn't clear if it
should be log based 10 or log base e

- replace the simple integrations using sum() with Simpson's rule
integrations.  this makes no visible difference in the final plots, but
might as well do this carefully since there's a function in scipy that does
it.

- remove redundant variables, simplify command line validation, etc.