ligolw_inspinjfind:  tie internal bisection search to --time-window
The internal bisection search for events near an injection was fixed to +/- 1 s, but when the --time-window option was create its default value was set to 9s, so many injections people were expecting to be able to find were not being found.  This patch ties the bisection search to the command line option.