glue.ligolw.utils:  use **kwargs in I/O functions
in particular, replace the xsl_file keyword argument of write_fileobj() with **kwargs.  by not passing this unconditionally to the .write() method of the xmldoc argument, any object with a .write() method can be passed to this function (and also to the other write_*() functions built on top of this).  this fixes a recently discovered regression whereby only genuine glue.ligolw.ligolw.Document objects may be written to disk with this code.