svim: handle zero zero-lag event case
This patch also introduces a flag that must be turned on to ensure
plotting of lambda fits. In general, this code should not be doing any
plotting. It was a mistake to let it do so. Instead, let svim do the
calculations it wants to do and output all data to file. Teach sink
how to read in the data, and let sink deal with the multitude of ways
that plotting inputs can go sour.