frutils.py: Update segments and frame lists in order
When updating the remote segment/file lists, these variables were being 'extend'd, which meant the fundamental segmentlist and frame list could be unordered. This expectation leads to strange errors when the index of the first desired frame is located. Bisect assumes the list is sorted and will produce incorrect results with unsorted lists of objects. That behavior leads to an incorrect index being picked out. So, as the time span was chunked up and queried against the (assumed ordered) frame list, the span of time and the frame list did not match and the underlying bindings were raising an error. The error was not as informative as it could have been (remedied by the previous patch).

The rememdy to this is to ensure that when the _remote* variables are updated, that they are inserted in order. All subsequent calls to bisect will behave properly. It is likely that this will degrade performance, but only noticeably for *very* large sets of frames/segments.