trig_combiner will correctly calculate trials
Previously, the code was hard coded to expect half of the first segment
to be lost, but this doesn't apply if we analyse the end of the segment
with the option --analyse-segment-end passed to
lalapps_coh_PTF_inspiral. Now the correct start time of the analysed
offsource is passed as a new option. Redundant options have been
removed.